Add garbage collector controller and move garbage collection of app resources
