+++ pkg/reconciler/app/reconciler.go
@@ -18,9 +18,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"math"
 	"reflect"
-	"sort"
 
 	"github.com/google/go-containerregistry/pkg/name"
 	containerregistryv1 "github.com/google/go-containerregistry/pkg/v1"
@@ -475,77 +473,6 @@ func (r *Reconciler) ApplyChanges(ctx context.Context, app *v1alpha1.App) error 
 		}
 	}
 
-	// GC'ing Builds
-	{
-		logger.Debug("GC'ing Builds")
-
-		configDefaults, err := kfconfig.FromContext(ctx).Defaults()
-		if err != nil {
-			return fmt.Errorf("failed to read config-defaults: %v", err)
-		}
-
-		buildLabelSelector := fmt.Sprintf("%s=%s", v1alpha1.NameLabel, app.Name)
-		listOptions := metav1.ListOptions{
-			LabelSelector: buildLabelSelector,
-		}
-		buildList, err := r.KfClientSet.
-			KfV1alpha1().
-			Builds(app.GetNamespace()).
-			List(ctx, listOptions)
-		if err != nil {
-			return err
-		}
-
-		if len(buildList.Items) > 0 {
-			maxBuildCount := v1alpha1.DefaultBuildRetentionCount
-			if configDefaults.BuildRetentionCount != nil {
-				maxBuildCount = int(*configDefaults.BuildRetentionCount)
-			}
-
-			buildsToDelete := buildsToGC(buildList.Items, maxBuildCount)
-			for _, t := range buildsToDelete {
-				if err := r.KfClientSet.KfV1alpha1().
-					Builds(app.GetNamespace()).
-					Delete(ctx, t.Name, metav1.DeleteOptions{}); err != nil {
-					return err
-				}
-			}
-		}
-	}
-
-	// reconcile Tasks
-	{
-		logger.Debug("reconciling tasks")
-		configDefaults, err := kfconfig.FromContext(ctx).Defaults()
-		if err != nil {
-			return fmt.Errorf("failed to read config-defaults: %v", err)
-		}
-		taskLabelSelector := fmt.Sprintf("%s=%s", v1alpha1.NameLabel, app.Name)
-		listOptions := metav1.ListOptions{
-			LabelSelector: taskLabelSelector,
-		}
-		taskList, err := r.KfClientSet.KfV1alpha1().Tasks(app.GetNamespace()).List(ctx, listOptions)
-		if err != nil {
-			return err
-		}
-
-		if len(taskList.Items) > 0 {
-			tasksToDelete := tasksToGC(taskList.Items, v1alpha1.DefaultMaxTaskCount)
-
-			if configDefaults.TaskRetentionCount != nil {
-				tasksToDelete = tasksToGC(taskList.Items, int(*configDefaults.TaskRetentionCount))
-			}
-
-			for _, t := range tasksToDelete {
-				if err := r.KfClientSet.KfV1alpha1().
-					Tasks(app.GetNamespace()).
-					Delete(ctx, t.Name, metav1.DeleteOptions{}); err != nil {
-					return err
-				}
-			}
-		}
-	}
-
 	instanceStatus := app.Spec.Instances.Status()
 
 	// reconcile HorizontalPodAutoscaler
@@ -783,64 +710,4 @@ func (r *Reconciler) updateStatus(ctx context.Context, desired *v1alpha1.App) (*
 	existing.Status = desired.Status
 
 	return r.KfClientSet.KfV1alpha1().Apps(existing.GetNamespace()).UpdateStatus(ctx, existing, metav1.UpdateOptions{})
-}
-
-func tasksToGC(tasks []v1alpha1.Task, maxTasks int) []v1alpha1.Task {
-	if len(tasks) <= maxTasks {
-		return nil
-	}
-
-	// Sort tasks first by status (final/non-final), then by creation
-	// timestamp (youngest first).  This ensures we GC the oldest completed
-	// Tasks.
-	sort.Slice(tasks, func(i, j int) bool {
-		taskA := tasks[i]
-		taskB := tasks[j]
-		taskACompleted := v1alpha1.IsStatusFinal(taskA.Status.Status)
-		taskBCompleted := v1alpha1.IsStatusFinal(taskB.Status.Status)
-
-		if taskACompleted == taskBCompleted {
-			return taskB.CreationTimestamp.Before(&taskA.CreationTimestamp)
-		}
-
-		return !taskACompleted
-	})
-
-	var tasksToGC []v1alpha1.Task
-
-	// Excess final tasks are deleted.
-	tasksOverLimit := tasks[maxTasks:]
-	for _, t := range tasksOverLimit {
-		if v1alpha1.IsStatusFinal(t.Status.Status) {
-			tasksToGC = append(tasksToGC, t)
-		}
-	}
-	return tasksToGC
-}
-
-func buildsToGC(builds []v1alpha1.Build, maxBuilds int) []v1alpha1.Build {
-	// Only GC final (Succeded=True/False) builds, non-final builds (Succeded=UNKNOWN)
-	// will fail after timeout (default 1 hour) and turn into Succeded=False.
-	var finalBuilds []v1alpha1.Build
-	for _, b := range builds {
-		if v1alpha1.IsStatusFinal(b.Status.Status) {
-			finalBuilds = append(finalBuilds, b)
-		}
-	}
-
-	// A minimum of one build should be kept -
-	// It continues to be needed for reference (e.g. "kf restage" re-runs the latest build)
-	maxBuilds = int(math.Max(1, float64(maxBuilds)))
-	if len(finalBuilds) <= maxBuilds {
-		return nil
-	}
-
-	// Sort builds by creation timestamp (youngest first).
-	// This ensures we GC the oldest completed Builds.
-	sort.Slice(finalBuilds, func(i, j int) bool {
-		return finalBuilds[j].CreationTimestamp.Before(&finalBuilds[i].CreationTimestamp)
-	})
-
-	// Excess final builds are deleted.
-	return finalBuilds[maxBuilds:]
 }
